+++ ui/packages/components/src/SQLEditor/types.ts
@@ -1,0 +1,6 @@
+export interface SQLCompletionConfig {
+  columns: readonly string[];
+  keywords: readonly string[];
+  functions: readonly { name: string; signature: string }[];
+  tables: readonly string[];
+}
